requirements.txt
pymysql
pillow
tensorflow
matplotlib
opencv-python
konlpy
schedule